removing duplicated 1
