+++ mathics/builtin/atomic/numbers.py
@@ -174,89 +174,6 @@ class ExactNumberQ(Test):
         return isinstance(expr, Number) and not expr.is_inexact()
 
 
-class IntegerDigits(Builtin):
-    """
-    <dl>
-    <dt>'IntegerDigits[$n$]'
-        <dd>returns a list of the base-10 digits in the integer $n$.
-    <dt>'IntegerDigits[$n$, $base$]'
-        <dd>returns a list of the base-$base$ digits in $n$.
-    <dt>'IntegerDigits[$n$, $base$, $length$]'
-        <dd>returns a list of length $length$, truncating or padding
-        with zeroes on the left as necessary.
-    </dl>
-
-    >> IntegerDigits[76543]
-     = {7, 6, 5, 4, 3}
-
-    The sign of $n$ is discarded:
-    >> IntegerDigits[-76543]
-     = {7, 6, 5, 4, 3}
-
-    >> IntegerDigits[15, 16]
-     = {15}
-    >> IntegerDigits[1234, 16]
-     = {4, 13, 2}
-    >> IntegerDigits[1234, 10, 5]
-     = {0, 1, 2, 3, 4}
-
-    #> IntegerDigits[1000, 10]
-     = {1, 0, 0, 0}
-
-    #> IntegerDigits[0]
-     = {0}
-    """
-
-    attributes = A_LISTABLE | A_PROTECTED
-
-    messages = {
-        "int": "Integer expected at position 1 in `1`",
-        "ibase": "Base `1` is not an integer greater than 1.",
-    }
-
-    rules = {
-        "IntegerDigits[n_]": "IntegerDigits[n, 10]",
-    }
-
-    summary_text = "digits of an integer in any base"
-
-    def apply_len(self, n, base, length, evaluation):
-        "IntegerDigits[n_, base_, length_]"
-
-        if not (isinstance(length, Integer) and length.get_int_value() >= 0):
-            return evaluation.message("IntegerDigits", "intnn")
-
-        return self.apply(n, base, evaluation, nr_elements=length.get_int_value())
-
-    def apply(self, n, base, evaluation, nr_elements=None):
-        "IntegerDigits[n_, base_]"
-
-        if not (isinstance(n, Integer)):
-            return evaluation.message(
-                "IntegerDigits", "int", Expression(SymbolIntegerDigits, n, base)
-            )
-
-        if not (isinstance(base, Integer) and base.get_int_value() > 1):
-            return evaluation.message("IntegerDigits", "ibase", base)
-
-        if nr_elements == 0:
-            # trivial case: we don't want any digits
-            return ListExpression()
-
-        # Note: above we checked that n and b are Integers, so we can use x.value.
-        digits = convert_int_to_digit_list(n.value, base.value)
-
-        if nr_elements is not None:
-            if len(digits) >= nr_elements:
-                # Truncate, preserving the digits on the right
-                digits = digits[-nr_elements:]
-            else:
-                # Pad with zeroes
-                digits = [0] * (nr_elements - len(digits)) + digits
-
-        return to_mathics_list(*digits, element_conversion_fn=Integer)
-
-
 class IntegerExponent(Builtin):
     """
     <dl>
@@ -796,8 +713,8 @@ class MaxPrecision(Predefined):
 class MachineEpsilon_(Predefined):
     """
     <dl>
-    <dt>'$MachineEpsilon'
-        <dd>is the distance between '1.0' and the next
+      <dt>'$MachineEpsilon'
+      <dd>is the distance between '1.0' and the next
             nearest representable machine-precision number.
     </dl>
 
@@ -923,8 +840,8 @@ class MinPrecision(Builtin):
 class NumericQ(Builtin):
     """
     <dl>
-    <dt>'NumericQ[$expr$]'
-        <dd>tests whether $expr$ represents a numeric quantity.
+      <dt>'NumericQ[$expr$]'
+      <dd>tests whether $expr$ represents a numeric quantity.
     </dl>
 
     >> NumericQ[2]
@@ -971,8 +888,8 @@ class NumericQ(Builtin):
 class Precision(Builtin):
     """
     <dl>
-    <dt>'Precision[$expr$]'
-        <dd>examines the number of significant digits of $expr$.
+      <dt>'Precision[$expr$]'
+      <dd>examines the number of significant digits of $expr$.
     </dl>
     This is rather a proof-of-concept than a full implementation.
     Precision of compound expression is not supported yet.
